requirements/conda_torch_cuda.txt
torchvision==0.11.1
pytorch==1.10.0
